+++ src/main/java/com/zysn/passwordmanager/model/security/algorithm/config/AlgorithmConfig.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.NoSuchElementException;
 
-import com.zysn.passwordmanager.model.utils.CryptoUtils;
+import com.zysn.passwordmanager.model.utils.crypto.CryptoUtils;
 
 /**
  * Algorithm configuration class for creating configuration for algorithm.
